[refact] HomeVC View 분리
